Moved DLLs
Moved DLLs inside "Sources" folder.
Call DLLs using .* in order to adapt to .so for linux.
All DLLs are in the same folder, for windows bitness is appended to the name. The goal is to "copy" the correct bitness ones during VIPM post install by removing the bitness in the name.